pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "bob/__init__.py"

[tool.aerich]
tortoise_orm = "bob.db.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."


[project]
name = "bob"
dynamic = ["version"]
dependencies = [
  "discord.py",
  "Levenshtein",
  "nest-asyncio",
  "jishaku",
  "pynacl",
  "topggpy",
  "yaml",
]

[project.scripts]
bob = "bob.__main__:main"
